Refactor, defaultBalance from DB, %not a number bug fix

+++ Assets/Scripts/PurchaseBuffBehaviour.cs
@@ -8,7 +8,7 @@ using UnityEngine.UI;
 
 public class PurchaseBuffBehaviour : MonoBehaviour
 {
-    private int _buffPrice;
+    private double _buffPrice;
     [SerializeField] private MoneyView moneyView;
     [SerializeField] private Button purchaseBuffButton;
